--- src/models/classes.js
@@ -1,18 +1,0 @@
-const mongoose = require('mongoose')
-
-const classSchema = new mongoose.Schema({
-    className: {
-      type: String,
-      required: true,
-    },
-    students: [
-      {
-        type: mongoose.Schema.Types.ObjectId,
-        ref: 'Student',
-      },
-    ],
-  });
-
-  const Class = mongoose.model('Class', classSchema)
-
-  module.exports = Class
