+++ src/models/teacherUser.js
@@ -1,0 +1,13 @@
+const mongoose = require('mongoose');
+
+const teacherUserSchema = new mongoose.Schema({
+    firstName: { type: String, required: true },
+    lastName: { type: String, required: true },
+    schoolName: { type: String, required: true },
+    username: { type: String, unique: true, required: true },
+    password: { type: String, required: true },
+  });
+
+const teacherUser = mongoose.model('teacherUser', teacherUserSchema);
+
+module.exports = teacherUser;
