--- src/models/users.js
@@ -1,29 +1,0 @@
-const mongoose = require('mongoose');
-
-const userSchema = new mongoose.Schema({
-    firstName: {
-        type: String,
-        required: true
-      },
-    lastName: {
-        type: String,
-        required: true
-      },
-    username: {
-        type: String,
-        required: true,
-        unique: true
-    },
-    password: {
-        type: String,
-        required: true
-    },
-    schoolName: {
-        type: String,
-        required: true
-    }
-});
-
-const User = mongoose.model('User', userSchema);
-
-module.exports = User;
